+++ src/views/AuthoritiesSearch/AuthoritiesSearch.js
@@ -2,6 +2,7 @@ import {
   useState,
   useEffect,
   useContext,
+  useMemo,
 } from 'react';
 import {
   useHistory,
@@ -74,7 +75,7 @@ const propTypes = {
   pageSize: PropTypes.number.isRequired,
   query: PropTypes.string,
   sortedColumn: PropTypes.string.isRequired,
-  sortOrder: PropTypes.oneOf([sortOrders.ASC, sortOrders.DESC]).isRequired,
+  sortOrder: PropTypes.oneOf([sortOrders.ASC, sortOrders.DES]).isRequired,
   totalRecords: PropTypes.number.isRequired,
 };
 
@@ -171,9 +172,10 @@ const AuthoritiesSearch = ({
   const [storedFilterPaneVisibility] = useLocalStorage(filterPaneVisibilityKey, true);
   const [isFilterPaneVisible, setIsFilterPaneVisible] = useState(storedFilterPaneVisibility);
   const [selectedRows, setSelectedRows] = useState({});
+  const [selectAll, setSelectAll] = useState(false);
 
   const selectedRowsIds = Object.keys(selectedRows);
-  const selectedRowsCount = selectedRowsIds.length;
+  const selectedRowsCount = useMemo(() => (Object.keys(selectedRows).length), [selectedRows]);
 
   const resetSelectedRows = () => {
     setSelectedRows({});
@@ -204,12 +206,50 @@ const AuthoritiesSearch = ({
     },
   });
 
-  const getNextSelectedRowsState = (prevSelectedRows, row) => {
-    const { id } = row;
-    const isRowSelected = !!prevSelectedRows[id];
-    const newSelectedRows = { ...prevSelectedRows };
+  const uniqueAuthoritiesCount = useMemo(() => {
+    // determine count of unique ids in authorities array.
+    // this is needed to check or uncheck "Select all" checkbox in header when all rows are explicitly
+    // checked or unchecked.
+    const filteredAuthorities = authorities.map(authority => authority.id).filter(id => !!id);
 
-    if (isRowSelected) {
+    return new Set(filteredAuthorities).size;
+  }, [authorities]);
+
+  const resetSelectedRows = () => {
+    setSelectedRows({});
+  };
+
+  const { exportRecords } = useAuthorityExport({
+    onError: () => {
+      const message = (
+        <FormattedMessage
+          id="ui-marc-authorities.export.failure"
+        />
+      );
+
+      callout.sendCallout({ type: 'error', message });
+    },
+    onSuccess: (data) => {
+      reportGenerator.toCSV(selectedRowsIds);
+
+      const message = (
+        <FormattedMessage
+          id="ui-marc-authorities.export.success"
+          values={{ exportJobName: data.jobExecutionId }}
+        />
+      );
+
+      callout.sendCallout({ type: 'success', message });
+      resetSelectedRows();
+    },
+  });
+
+  const getNextSelectedRowsState = (row) => {
+    const { id } = row;
+    const isRowSelected = !!selectedRows[id];
+    const newSelectedRows = { ...selectedRows };
+
+    if (isRowSelected || selectAll) {
       delete newSelectedRows[id];
     } else {
       newSelectedRows[id] = row;
@@ -218,8 +258,34 @@ const AuthoritiesSearch = ({
     return newSelectedRows;
   };
 
+  const getSelectAllRowsState = () => {
+    if (!selectAll) {
+      return authorities.filter(item => !!item.id).reduce((acc, item) => {
+        return {
+          ...acc,
+          [item.id]: item,
+        };
+      }, {});
+    } else {
+      return {};
+    }
+  };
+
   const toggleRowSelection = (row) => {
-    setSelectedRows(prev => getNextSelectedRowsState(prev, row));
+    const newRows = getNextSelectedRowsState(row);
+
+    setSelectedRows(newRows);
+    if (
+      (Object.keys(newRows).length === uniqueAuthoritiesCount && !selectAll) ||
+      (Object.keys(newRows).length === 0 && selectAll)
+    ) {
+      setSelectAll(prev => !prev);
+    }
+  };
+
+  const toggleSelectAll = () => {
+    setSelectedRows(getSelectAllRowsState());
+    setSelectAll(prev => !prev);
   };
 
   const toggleFilterPane = () => {
@@ -390,6 +456,8 @@ const AuthoritiesSearch = ({
           isFilterPaneVisible={isFilterPaneVisible}
           toggleFilterPane={toggleFilterPane}
           toggleRowSelection={toggleRowSelection}
+          toggleSelectAll={toggleSelectAll}
+          selectAll={selectAll}
           hasFilters={!!filters.length}
           query={searchQuery}
           hidePageIndices={hidePageIndices}